Convert Python integration to provider arrays
